add func_percent_rank_test.go
